Fixed Segmentation Fault when CMD list is empty
When one brings up a command palette without any commands registered,
and press enter. The application will segfault.

I have identified and fixed the place where this happen.
I've added a check to see if the search is empty or when the search
result is empty, then it does nothing.